Refactor update_state among runners.
